rewrite fillstored!(::SpecialMatrix, x) without generated function move implementations to more suitable places
